This is the framework for an AI project I've used with AP CS students.
How do you play Tic-Tac-Toe?  Students will program an AI to play the
game of Tic-Tac-Toe.  The goal is to create an AI to play other AIs
built by classmates.  The program can simulate thousands of games and
give a result.

Once the class has all created AIs, there is a tournament function to
output the results to a CSV file.